authenticate post creation and update

+++ app/Http/Controllers/API/PostController.php
@@ -5,8 +5,10 @@ namespace App\Http\Controllers\API;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Post;
+use App\User;
 use App\Http\Resources\PostResource;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -29,38 +31,48 @@ class PostController extends Controller
 
     public function show($id){
         return new PostResource(Post::find($id));
-        // return Post::find($id);
     }
 
+    //only authenticated user
     public function store(Request $request){
         if ($request->profile){
             $request['image']=Storage::disk('public')->put('img',$request->profile);
         }
-        $request['user_id'] =1; //untill i have a user
+        $user=Auth::user();
+        $request['user_id'] =$user->id;
         $post=Post::create($request->all());
         return response()->json($post,201);
     }
 
+    //only authenticated user
     public function update(Request $request, $id){
-        $request['user_id'] =1;//untill i have a user
+        // $user=Auth::user();
+        // $request['user_id'] =$user->id; 
         $updated_post=Post::find($id);
         $updated_post->update($request->all());
         $updated_post->fresh(); #fresh return new instance of the model
         return response()->json($updated_post, 200);
     }
 
+    //only authenticated user
     public function destroy(Request $request, $id){
-        $request['user_id'] =1;
+       if (Auth::user()){
         $deleted_post=Post::find($id);
         $del=$deleted_post->delete();
         if ($del==1){
-            // $success["message"] = "Post Deleted Successfully";
-            return response()->json($deleted_post);
-        }else{
-            return "error";
+            $success["message"] = "Post Deleted Successfully";
+            return response()->json($success);
         }
-        // $success["message"] = "Post Deleted Successfully";
-        // return $success["message"];
+        // }else{
+        //     $error["message"] =" Cannot Delete This Post";
+        //     return response()->json($error);
+        // }
+
+       }else{
+        $error["message"] =" unautheried";
+        return response()->json($error); 
+       }
+        
     }
